refactor(server): split api.rs into sessions / git / system submodules
Cleanup-pass finding #2 from the code-quality scan. No behavior changes;
all 1,213 lib tests pass with --features serve.

Before: src/server/api.rs was 2,151 lines hosting 28 axum handlers and
their request/response DTOs across five unrelated domains.

After:
  - src/server/api/mod.rs (49 lines) — module declarations, re-exports
    of every public handler so external callers keep `api::*` paths,
    and the small set of shared validation helpers.
  - src/server/api/sessions.rs (1,485 lines) — session CRUD, the
    ensure-* lifecycle endpoints, and the per-file rich-diff handlers
    (kept together because diff is per-session). Includes the
    SessionResponse and validate_diff_path tests.
  - src/server/api/git.rs (293 lines) — clone_repo and list_branches
    plus the repo_name_from_url tests.
  - src/server/api/system.rs (384 lines) — agents, settings, themes,
    profiles, filesystem browse, groups, docker status, devices, about.

Server router in src/server/mod.rs is unchanged because every handler
keeps its `api::handler` path through the re-exports.
