+++ src/server/api/sessions.rs
@@ -1,4 +1,4 @@
-//! REST API handlers for session management and agents.
+//! Session CRUD, ensure-* lifecycle endpoints, and per-file diff handlers.
 
 use std::sync::Arc;
 
@@ -13,31 +13,8 @@ use serde::{Deserialize, Serialize};
 use crate::session::{Instance, Status, Storage};
 
 use super::AppState;
+use super::validate_no_shell_injection;
 
-const SHELL_METACHARACTERS: &[char] = &[
-    ';', '&', '|', '$', '`', '(', ')', '{', '}', '<', '>', '\n', '\r', '\\', '"', '\'', '!', '#',
-    '*', '?', '[', ']', '~', '\t', '\0',
-];
-
-fn validate_no_shell_injection(value: &str, field_name: &str) -> Result<(), String> {
-    if let Some(c) = value.chars().find(|c| SHELL_METACHARACTERS.contains(c)) {
-        return Err(format!(
-            "Invalid character '{}' in {}. Shell metacharacters are not allowed.",
-            c, field_name
-        ));
-    }
-    Ok(())
-}
-
-const ALLOWED_SETTINGS_SECTIONS: &[&str] = &[
-    "theme", "session", "tmux", "updates", "sound", "sandbox", "worktree",
-];
-
-const SESSION_BLOCKED_FIELDS: &[&str] =
-    &["agent_command_override", "agent_extra_args", "extra_env"];
-
-/// API response DTO for session data.
-/// Decouples the API contract from the internal Instance struct.
 #[derive(Serialize)]
 pub struct SessionResponse {
     pub id: String,
@@ -145,7 +122,7 @@ pub async fn list_sessions(State(state): State<Arc<AppState>>) -> Json<Vec<Sessi
                     })
             });
         }
-        *state.cleanup_defaults_cache.write().await = super::CleanupDefaultsCache {
+        *state.cleanup_defaults_cache.write().await = crate::server::CleanupDefaultsCache {
             refreshed_at: std::time::Instant::now(),
             entries: fresh.clone(),
         };
@@ -680,174 +657,6 @@ pub async fn create_session(
         }
         Err(e) => {
             tracing::error!("Session creation panicked: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(serde_json::json!({"error": "internal", "message": "Internal server error"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-// --- Clone repository ---
-
-#[derive(Deserialize)]
-pub struct CloneRepoBody {
-    pub url: String,
-    pub destination: Option<String>,
-    #[serde(default)]
-    pub shallow: bool,
-}
-
-/// Extract a repository name from a git URL.
-///
-/// Handles HTTPS, SSH, and scp-style URLs:
-///   https://github.com/user/repo.git       -> repo
-///   git@github.com:user/repo.git           -> repo
-///   ssh://git@host/user/repo               -> repo
-///   ssh://git@host:2222/user/repo.git      -> repo
-fn repo_name_from_url(url: &str) -> Option<String> {
-    // For scheme-based URLs (https://, ssh://, git://), take the last path segment.
-    // For scp-style (git@host:path), split on ':' and take the last segment of the path.
-    let last_segment = if url.contains("://") {
-        url.rsplit_once('/')?.1
-    } else if let Some((_host, path)) = url.split_once(':') {
-        // scp-style: git@github.com:user/repo.git
-        path.rsplit('/').next().unwrap_or(path)
-    } else {
-        url.rsplit_once('/')?.1
-    };
-    let name = last_segment.trim_end_matches(".git");
-    if name.is_empty() {
-        None
-    } else {
-        Some(name.to_string())
-    }
-}
-
-pub async fn clone_repo(
-    State(state): State<Arc<AppState>>,
-    Json(body): Json<CloneRepoBody>,
-) -> impl IntoResponse {
-    if state.read_only {
-        return (
-            StatusCode::FORBIDDEN,
-            Json(
-                serde_json::json!({"error": "read_only", "message": "Server is in read-only mode"}),
-            ),
-        )
-            .into_response();
-    }
-
-    let url = body.url.trim().to_string();
-    if url.is_empty() {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(
-                serde_json::json!({"error": "validation_failed", "message": "URL cannot be empty"}),
-            ),
-        )
-            .into_response();
-    }
-
-    // Basic URL scheme validation
-    let looks_like_url = url.starts_with("https://")
-        || url.starts_with("http://")
-        || url.starts_with("git://")
-        || url.starts_with("ssh://")
-        || (url.contains('@') && url.contains(':') && !url.contains(' '));
-    if !looks_like_url {
-        return (
-            StatusCode::BAD_REQUEST,
-            Json(serde_json::json!({"error": "validation_failed", "message": "URL does not look like a git repository URL"})),
-        )
-            .into_response();
-    }
-
-    // Resolve destination path
-    let destination = if let Some(ref dest) = body.destination {
-        let dest = dest.trim();
-        if dest.is_empty() {
-            None
-        } else {
-            Some(std::path::PathBuf::from(dest))
-        }
-    } else {
-        None
-    };
-
-    let destination = match destination {
-        Some(d) => d,
-        None => {
-            let repo_name = repo_name_from_url(&url).unwrap_or_else(|| "cloned-repo".to_string());
-            let home = dirs::home_dir().unwrap_or_else(|| std::path::PathBuf::from("/tmp"));
-            home.join(&repo_name)
-        }
-    };
-
-    // Security: destination must be within the home directory
-    if let Some(home) = dirs::home_dir() {
-        let canonical_home = home.canonicalize().unwrap_or(home);
-        // For new paths that don't exist yet, check the parent
-        let check_path = if destination.exists() {
-            destination.canonicalize().unwrap_or(destination.clone())
-        } else {
-            destination
-                .parent()
-                .and_then(|p| p.canonicalize().ok())
-                .map(|p| p.join(destination.file_name().unwrap_or_default()))
-                .unwrap_or(destination.clone())
-        };
-        if !check_path.starts_with(&canonical_home) {
-            return (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({"error": "validation_failed", "message": "Destination must be within the home directory"})),
-            )
-                .into_response();
-        }
-    }
-
-    let dest_display = destination.display().to_string();
-
-    // Return an actionable error if the destination already exists, before
-    // spawning the blocking task, so the user knows to pick a different name.
-    if destination.exists() {
-        return (
-            StatusCode::CONFLICT,
-            Json(serde_json::json!({
-                "error": "destination_exists",
-                "message": format!(
-                    "Directory '{}' already exists. Choose a different destination or delete it first.",
-                    destination.file_name().unwrap_or_default().to_string_lossy()
-                ),
-                "path": dest_display,
-            })),
-        )
-            .into_response();
-    }
-
-    let shallow = body.shallow;
-    let result = tokio::task::spawn_blocking(move || {
-        crate::git::clone_repo(&url, &destination, shallow)?;
-        Ok::<String, crate::git::error::GitError>(destination.display().to_string())
-    })
-    .await;
-
-    match result {
-        Ok(Ok(path)) => {
-            (StatusCode::CREATED, Json(serde_json::json!({"path": path}))).into_response()
-        }
-        Ok(Err(e)) => {
-            let msg = e.to_string();
-            tracing::warn!("Clone failed for {dest_display}: {msg}");
-            (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({"error": "clone_failed", "message": msg})),
-            )
-                .into_response()
-        }
-        Err(e) => {
-            tracing::error!("Clone task panicked: {e}");
             (
                 StatusCode::INTERNAL_SERVER_ERROR,
                 Json(serde_json::json!({"error": "internal", "message": "Internal server error"})),
@@ -1493,442 +1302,10 @@ pub async fn session_diff_file(
     }
 }
 
-// --- Agents ---
-
-#[derive(Serialize)]
-pub struct AgentInfo {
-    pub name: String,
-    pub binary: String,
-    pub host_only: bool,
-    pub installed: bool,
-}
-
-pub async fn list_agents() -> Json<Vec<AgentInfo>> {
-    let result = tokio::task::spawn_blocking(|| {
-        let tools = crate::tmux::AvailableTools::detect();
-        let available = tools.available_list();
-        crate::agents::AGENTS
-            .iter()
-            .map(|a| AgentInfo {
-                name: a.name.to_string(),
-                binary: a.binary.to_string(),
-                host_only: a.host_only,
-                installed: available.iter().any(|s| s == a.name),
-            })
-            .collect::<Vec<_>>()
-    })
-    .await
-    .unwrap_or_default();
-    Json(result)
-}
-
-// --- Settings ---
-
-#[derive(Deserialize)]
-pub struct SettingsQuery {
-    pub profile: Option<String>,
-}
-
-pub async fn get_settings(
-    axum::extract::Query(query): axum::extract::Query<SettingsQuery>,
-) -> impl IntoResponse {
-    let config_result = if let Some(ref profile_name) = query.profile {
-        crate::session::resolve_config(profile_name)
-    } else {
-        crate::session::Config::load()
-    };
-
-    match config_result {
-        Ok(config) => match serde_json::to_value(&config) {
-            Ok(val) => (StatusCode::OK, Json(val)).into_response(),
-            Err(e) => {
-                tracing::error!("Settings serialization failed: {}", e);
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(serde_json::json!({"error": "serialize_failed", "message": "Failed to serialize settings"})),
-                )
-                    .into_response()
-            }
-        },
-        Err(e) => {
-            tracing::error!("Settings load failed: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(serde_json::json!({"error": "load_failed", "message": "Failed to load settings"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-pub async fn update_settings(Json(body): Json<serde_json::Value>) -> impl IntoResponse {
-    // Validate that only allowed sections are being updated
-    if let Some(obj) = body.as_object() {
-        for key in obj.keys() {
-            if !ALLOWED_SETTINGS_SECTIONS.contains(&key.as_str()) {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(serde_json::json!({
-                        "error": "validation_failed",
-                        "message": format!("Settings section '{}' is not allowed via the web API.", key)
-                    })),
-                )
-                    .into_response();
-            }
-        }
-    }
-
-    let result = tokio::task::spawn_blocking(move || {
-        let config = crate::session::Config::load().unwrap_or_default();
-        let mut current = serde_json::to_value(&config)?;
-        if let (Some(current_obj), Some(update_obj)) = (current.as_object_mut(), body.as_object()) {
-            for (key, value) in update_obj {
-                let mut value = value.clone();
-                // Strip blocked fields from session section
-                if key == "session" {
-                    if let Some(session_obj) = value.as_object_mut() {
-                        for blocked in SESSION_BLOCKED_FIELDS {
-                            session_obj.remove(*blocked);
-                        }
-                    }
-                }
-                current_obj.insert(key.clone(), value);
-            }
-        }
-        let config: crate::session::Config = serde_json::from_value(current)?;
-        crate::session::save_config(&config)?;
-        Ok::<_, anyhow::Error>(config)
-    })
-    .await;
-
-    match result {
-        Ok(Ok(config)) => match serde_json::to_value(&config) {
-            Ok(val) => (StatusCode::OK, Json(val)).into_response(),
-            Err(e) => {
-                tracing::error!("Settings serialization failed: {}", e);
-                (
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(serde_json::json!({"error": "serialize_failed", "message": "Failed to serialize settings"})),
-                )
-                    .into_response()
-            }
-        },
-        Ok(Err(e)) => {
-            tracing::warn!("Settings update failed: {}", e);
-            (
-                StatusCode::BAD_REQUEST,
-                Json(serde_json::json!({"error": "update_failed", "message": "Failed to update settings"})),
-            )
-                .into_response()
-        }
-        Err(e) => {
-            tracing::error!("Settings update panicked: {}", e);
-            (
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(serde_json::json!({"error": "internal", "message": "Internal server error"})),
-            )
-                .into_response()
-        }
-    }
-}
-
-// --- Devices ---
-
-pub async fn list_devices(State(state): State<Arc<AppState>>) -> Json<Vec<super::DeviceInfo>> {
-    let devices = state.devices.read().await;
-    Json(devices.clone())
-}
-
-// --- Themes ---
-
-pub async fn list_themes() -> Json<Vec<String>> {
-    Json(
-        crate::tui::styles::available_themes()
-            .into_iter()
-            .map(|s| s.to_string())
-            .collect(),
-    )
-}
-
-// --- Wizard support ---
-
-#[derive(Serialize)]
-pub struct ProfileInfo {
-    pub name: String,
-    pub is_default: bool,
-}
-
-pub async fn list_profiles(State(state): State<Arc<AppState>>) -> Json<Vec<ProfileInfo>> {
-    let profiles = crate::session::list_profiles().unwrap_or_default();
-    let active = &state.profile;
-    let mut result: Vec<ProfileInfo> = profiles
-        .into_iter()
-        .map(|name| {
-            let is_default = name == *active;
-            ProfileInfo { name, is_default }
-        })
-        .collect();
-    // Ensure the active profile appears even if list_profiles missed it
-    if !result.iter().any(|p| p.name == *active) {
-        result.insert(
-            0,
-            ProfileInfo {
-                name: active.clone(),
-                is_default: true,
-            },
-        );
-    }
-    Json(result)
-}
-
-#[derive(Deserialize)]
-pub struct BrowseQuery {
-    pub path: String,
-    pub limit: Option<usize>,
-}
-
-#[derive(Serialize)]
-pub struct DirEntry {
-    pub name: String,
-    pub path: String,
-    pub is_dir: bool,
-    pub is_git_repo: bool,
-}
-
-#[derive(Serialize)]
-struct BrowseResponse {
-    entries: Vec<DirEntry>,
-    has_more: bool,
-}
-
-pub async fn filesystem_home() -> impl IntoResponse {
-    match dirs::home_dir() {
-        Some(home) => (
-            StatusCode::OK,
-            Json(serde_json::json!({"path": home.to_string_lossy()})),
-        )
-            .into_response(),
-        None => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(serde_json::json!({"error": "Could not determine home directory"})),
-        )
-            .into_response(),
-    }
-}
-
-pub async fn browse_filesystem(
-    axum::extract::Query(query): axum::extract::Query<BrowseQuery>,
-) -> impl IntoResponse {
-    let result = tokio::task::spawn_blocking(move || {
-        let limit = query.limit.unwrap_or(100);
-        let path = std::path::Path::new(&query.path);
-        let canonical = path.canonicalize().map_err(|_| "Path does not exist")?;
-
-        if !canonical.is_dir() {
-            return Err("Path is not a directory");
-        }
-
-        // Security: restrict browsing to the user's home directory
-        if let Some(home) = dirs::home_dir() {
-            if !canonical.starts_with(&home) {
-                return Err("Path is outside the home directory");
-            }
-        }
-
-        let mut entries: Vec<DirEntry> = Vec::new();
-        let read_dir = std::fs::read_dir(&canonical).map_err(|_| "Cannot read directory")?;
-
-        for entry in read_dir.flatten() {
-            let name = entry.file_name().to_string_lossy().to_string();
-            if name.starts_with('.') {
-                continue;
-            }
-            let entry_path = entry.path();
-            let is_dir = entry_path.is_dir();
-            if !is_dir {
-                continue;
-            }
-            let is_git_repo = entry_path.join(".git").exists();
-            entries.push(DirEntry {
-                name,
-                path: entry_path.to_string_lossy().to_string(),
-                is_dir,
-                is_git_repo,
-            });
-        }
-        entries.sort_by(|a, b| a.name.to_lowercase().cmp(&b.name.to_lowercase()));
-        let has_more = entries.len() > limit;
-        entries.truncate(limit);
-        Ok(BrowseResponse { entries, has_more })
-    })
-    .await;
-
-    match result {
-        Ok(Ok(resp)) => (StatusCode::OK, Json(serde_json::to_value(resp).unwrap())).into_response(),
-        Ok(Err(msg)) => (
-            StatusCode::BAD_REQUEST,
-            Json(serde_json::json!({"error": "browse_failed", "message": msg})),
-        )
-            .into_response(),
-        Err(e) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(serde_json::json!({"error": "internal", "message": e.to_string()})),
-        )
-            .into_response(),
-    }
-}
-
-#[derive(Deserialize)]
-pub struct BranchesQuery {
-    pub path: String,
-}
-
-#[derive(Serialize)]
-pub struct BranchInfo {
-    pub name: String,
-    pub is_current: bool,
-}
-
-pub async fn list_branches(
-    axum::extract::Query(query): axum::extract::Query<BranchesQuery>,
-) -> impl IntoResponse {
-    let result = tokio::task::spawn_blocking(move || {
-        let path = std::path::Path::new(&query.path);
-        if !crate::git::GitWorktree::is_git_repo(path) {
-            return Err("Path is not a git repository".to_string());
-        }
-
-        let branches = crate::git::diff::list_branches(path).map_err(|e| e.to_string())?;
-
-        let current = std::process::Command::new("git")
-            .args(["branch", "--show-current"])
-            .current_dir(path)
-            .output()
-            .ok()
-            .and_then(|o| String::from_utf8(o.stdout).ok())
-            .map(|s| s.trim().to_string())
-            .unwrap_or_default();
-
-        let mut result: Vec<BranchInfo> = branches
-            .into_iter()
-            .take(200)
-            .map(|name| {
-                let is_current = name == current;
-                BranchInfo { name, is_current }
-            })
-            .collect();
-
-        result.sort_by(|a, b| b.is_current.cmp(&a.is_current));
-
-        Ok(result)
-    })
-    .await;
-
-    match result {
-        Ok(Ok(branches)) => (
-            StatusCode::OK,
-            Json(serde_json::to_value(branches).unwrap()),
-        )
-            .into_response(),
-        Ok(Err(msg)) => (
-            StatusCode::BAD_REQUEST,
-            Json(serde_json::json!({"error": "not_a_repo", "message": msg})),
-        )
-            .into_response(),
-        Err(e) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(serde_json::json!({"error": "internal", "message": e.to_string()})),
-        )
-            .into_response(),
-    }
-}
-
-#[derive(Serialize)]
-pub struct GroupInfo {
-    pub path: String,
-    pub session_count: usize,
-}
-
-pub async fn list_groups(State(state): State<Arc<AppState>>) -> impl IntoResponse {
-    let instances = state.instances.read().await;
-    let mut group_counts: std::collections::HashMap<String, usize> =
-        std::collections::HashMap::new();
-    for inst in instances.iter() {
-        if !inst.group_path.is_empty() {
-            *group_counts.entry(inst.group_path.clone()).or_default() += 1;
-        }
-    }
-    let groups: Vec<GroupInfo> = group_counts
-        .into_iter()
-        .map(|(path, session_count)| GroupInfo {
-            path,
-            session_count,
-        })
-        .collect();
-    Json(groups)
-}
-
-#[derive(Serialize)]
-pub struct DockerStatus {
-    pub available: bool,
-    pub runtime: Option<String>,
-}
-
-pub async fn docker_status() -> Json<DockerStatus> {
-    let result = tokio::task::spawn_blocking(|| {
-        use crate::containers::ContainerRuntimeInterface;
-        let runtime = crate::containers::get_container_runtime();
-        let available = runtime.is_available() && runtime.is_daemon_running();
-        let runtime_name = if available {
-            let config = crate::session::Config::load().unwrap_or_default();
-            Some(
-                serde_json::to_value(config.sandbox.container_runtime)
-                    .ok()
-                    .and_then(|v| v.as_str().map(String::from))
-                    .unwrap_or_else(|| "docker".to_string()),
-            )
-        } else {
-            None
-        };
-        DockerStatus {
-            available,
-            runtime: runtime_name,
-        }
-    })
-    .await
-    .unwrap_or(DockerStatus {
-        available: false,
-        runtime: None,
-    });
-    Json(result)
-}
-
-#[derive(Serialize)]
-pub struct ServerAbout {
-    pub version: String,
-    pub auth_required: bool,
-    pub passphrase_enabled: bool,
-    pub read_only: bool,
-    pub behind_tunnel: bool,
-    pub profile: String,
-}
-
-pub async fn get_about(State(state): State<Arc<AppState>>) -> Json<ServerAbout> {
-    let auth_required = !state.token_manager.is_no_auth().await;
-    Json(ServerAbout {
-        version: env!("CARGO_PKG_VERSION").to_string(),
-        auth_required,
-        passphrase_enabled: state.login_manager.is_enabled(),
-        read_only: state.read_only,
-        behind_tunnel: state.behind_tunnel,
-        profile: state.profile.clone(),
-    })
-}
 
 #[cfg(test)]
 mod tests {
     use super::*;
-
     fn make_test_instance() -> Instance {
         let mut inst = Instance::new("test-session", "/tmp/test-project");
         inst.tool = "claude".to_string();
@@ -1996,7 +1373,6 @@ mod tests {
         assert_eq!(json["status"], "Running");
         assert_eq!(json["is_sandboxed"], false);
     }
-
     // ── validate_diff_path: security regression tests ──────────────────────────
     //
     // Regression for a path-traversal vulnerability in the first cut of the
@@ -2105,47 +1481,5 @@ mod tests {
             &changed(&["deleted.txt"]),
         );
         assert!(ok.is_ok(), "expected Ok, got {:?}", ok);
-    }
-
-    // ── repo_name_from_url tests ─────────────────────────────────────────────
-
-    #[test]
-    fn repo_name_from_https_url() {
-        assert_eq!(
-            repo_name_from_url("https://github.com/user/my-repo.git"),
-            Some("my-repo".to_string())
-        );
-    }
-
-    #[test]
-    fn repo_name_from_https_url_no_dotgit() {
-        assert_eq!(
-            repo_name_from_url("https://github.com/user/my-repo"),
-            Some("my-repo".to_string())
-        );
-    }
-
-    #[test]
-    fn repo_name_from_scp_url() {
-        assert_eq!(
-            repo_name_from_url("git@github.com:user/my-repo.git"),
-            Some("my-repo".to_string())
-        );
-    }
-
-    #[test]
-    fn repo_name_from_ssh_url() {
-        assert_eq!(
-            repo_name_from_url("ssh://git@github.com/user/my-repo.git"),
-            Some("my-repo".to_string())
-        );
-    }
-
-    #[test]
-    fn repo_name_from_ssh_url_with_port() {
-        assert_eq!(
-            repo_name_from_url("ssh://git@host:2222/user/my-repo.git"),
-            Some("my-repo".to_string())
-        );
     }
 }
